+++ robozonky-app/src/test/java/com/github/robozonky/app/delinquencies/TransactionalTenantTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.github.robozonky.app.daemon;
+package com.github.robozonky.app.delinquencies;
 
 import java.io.IOException;
 import java.util.function.Function;
